chat_with_human/requirements.txt
google-adk
requests
fastapi
uvicorn
google-generativeai